Progreso en hardware real: getCurrentLanguage, logs por ejecucion, apk/obb minimos
- java.c: stub de getCurrentLanguage (faltaba, coincide con la pantalla de menu
  sin texto vista en el log de hardware real).
- main.c: fixea warnings de tipo en nativeTouches* (casts a jfloat).
- logger.c: cada corrida ahora escribe a un archivo nuevo
  ux0:data/popclassic/logs/log_<unix_timestamp>_.txt (crea la carpeta logs/ si
  no existe) en vez de un log.txt unico acumulativo, para poder documentar el
  avance de cada prueba por separado.
- build_and_install.sh: tambien exporta eboot.bin a build/, no solo el vpk.
- plan_portabilidad.md (§9.19): documenta el diagnostico del Data abort en
  hardware real (original.apk faltante), y la investigacion de que el .apk/.obb
  originales (48MB/186MB) solo aportan assets/appConfig.txt y Localization/*.loc
  en runtime -- el resto ya se lee suelto. Deja registrado el plan para
  eventualmente hookear CCFileUtils::getFileData y sacarse de encima los ZIPs
  del todo, con el riesgo de allocator mismatch a resolver antes de activarlo.

+++ source/main.c
@@ -75,6 +75,7 @@ int main() {
         //   at the .apk, not at a folder.
         // Data/* loose assets under DATA_PATH are unaffected either way, since
         // those are read via plain fopen(), not through either of these.
+
         if (!file_exists(DATA_PATH "original.apk")) {
             fatal_error("Error: " DATA_PATH "original.apk not found.\n"
                         "Copy the original game's .apk there (renamed to "
@@ -127,16 +128,16 @@ int main() {
                 int y = (int)((float)touch.report[i].y * 544.0f / 1088.0f);
 
                 if (lastX[i] == -1 || lastY[i] == -1) {
-                    if (nativeTouchesBegin) nativeTouchesBegin(jniEnv, NULL, touch.report[i].id, x, y);
+                    if (nativeTouchesBegin) nativeTouchesBegin(jniEnv, NULL, touch.report[i].id, (jfloat)x, (jfloat)y);
                 } else {
                     if (lastX[i] != x || lastY[i] != y)
-                        if (nativeTouchesMove) nativeTouchesMove(jniEnv, NULL, touch.report[i].id, x, y);
+                        if (nativeTouchesMove) nativeTouchesMove(jniEnv, NULL, touch.report[i].id, (jfloat)x, (jfloat)y);
                 }
                 lastX[i] = x;
                 lastY[i] = y;
             } else {
                 if (lastX[i] != -1 || lastY[i] != -1) {
-                    if (nativeTouchesEnd) nativeTouchesEnd(jniEnv, NULL, i, lastX[i], lastY[i]);
+                    if (nativeTouchesEnd) nativeTouchesEnd(jniEnv, NULL, i, (jfloat)lastX[i], (jfloat)lastY[i]);
                     lastX[i] = -1;
                     lastY[i] = -1;
                 }
